Remove rule cache for now due to critical issues
